dev-requirements.txt
mypy ~= 1.10.0
pytest ~= 8.2.0
ruff ~= 0.4.2